pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docmirror-cli"
version = "0.1.0"
description = "Scaffold for a CLI that converts structured JPG/JPEG forms into mirrored editable DOCX."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12,<1.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0",
  "python-docx>=1.1",
  "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
analysis = [
  "docling>=2.20.0",
]
fallback = [
  "opencv-python-headless>=4.10.0.84",
  "pytesseract>=0.3.10",
  "lxml>=5.2",
  "Pillow>=10.0",
]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-PyYAML>=6.0.12",
  "tomli>=2.0",
]

[project.scripts]
docmirror = "docmirror.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.ruff.lint.per-file-ignores]
"src/docmirror/cli.py" = ["B008"]
"scripts/export_previous_response_report.py" = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_equality = true
mypy_path = ["src"]
